+++ src/pan_tilt_node.cpp
@@ -37,13 +37,14 @@ PanTiltNode::PanTiltNode()
 	n_.param("/servo/index1/tilt_max", tilt_max_[1], tilt_max_[1]);
 	n_.param("/servo/index1/tilt_min", tilt_min_[1], tilt_min_[1]);
 
+    first_index0_msg_received_ = false;
+    first_index1_msg_received_ = false;
+
     pan_tilt_sub_[0] = n_.subscribe("pan_tilt_node/index0_position", 10, &PanTiltNode::panTilt0CB, this);
 	pan_tilt_sub_[1] = n_.subscribe("pan_tilt_node/index1_position", 10, &PanTiltNode::panTilt1CB, this);
 
     // Published topic is latched
 	servo_array_pub_ = n_.advertise<servo_msgs::servo_array>("servo", 10, true);
-
-    first_reconfigure_ = true;
 }
 
 // This callback is for when the dynamic configuration parameters change
@@ -54,18 +55,17 @@ void PanTiltNode::reconfCallback(pan_tilt::PanTiltConfig &config, uint32_t level
     index1_pan_trim_ = config.index1_pan_trim;
     index1_tilt_trim_ = config.index1_tilt_trim;
 
-    // We don't want the first call which initalises the parameters
-    // to result in sending a message to the servos as the angle will
-    // be just the trim value if no position messages received yet
-    if(first_reconfigure_ == false)
+    // We don't want to send a message following a call here unless we have received
+    // a position message. Otherwise the trim value will be taken for an actual position.
+    if(first_index0_msg_received_ == true)
     {
         // Send new messages with new trim values
-        movePanTilt(index0_pan_tilt_, index0_pan_trim_, index0_tilt_trim_, 0);
-        movePanTilt(index1_pan_tilt_, index1_pan_trim_, index1_tilt_trim_, 1);
+        movePanTilt(index0_pan_tilt_, index0_pan_trim_, index0_tilt_trim_, 0);        
     }
-    else
+
+    if(first_index1_msg_received_ == true)
     {
-        first_reconfigure_ = false;
+        movePanTilt(index1_pan_tilt_, index1_pan_trim_, index1_tilt_trim_, 1);
     }
 }
 //---------------------------------------------------------------------------
@@ -73,8 +73,9 @@ void PanTiltNode::reconfCallback(pan_tilt::PanTiltConfig &config, uint32_t level
 // Callback to move the pan tilt device indexed 0
 void PanTiltNode::panTilt0CB(const servo_msgs::pan_tilt& pan_tilt)
 {
-    // Store lastes value in case we change thte trim
+    // Store latest value in case we change thte trim
     index0_pan_tilt_ = pan_tilt;
+    first_index0_msg_received_ = true;
     
     movePanTilt(pan_tilt, index0_pan_trim_, index0_tilt_trim_, 0);
 }
@@ -82,8 +83,9 @@ void PanTiltNode::panTilt0CB(const servo_msgs::pan_tilt& pan_tilt)
 // Callback to move the pan tilt device indexed 1
 void PanTiltNode::panTilt1CB(const servo_msgs::pan_tilt& pan_tilt)
 {    
-    // Store lastes value in case we change thte trim
+    // Store latest value in case we change thte trim
     index1_pan_tilt_ = pan_tilt;
+    first_index1_msg_received_ = true;
     
     movePanTilt(pan_tilt, index1_pan_trim_, index0_tilt_trim_, 1);
 }